pyproject.toml
[project]
name = "mcp-oauth-server"
version = "1.0.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "supabase>=2.0.0",
    "bcrypt>=4.0.0",
    "python-dotenv>=1.0.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.9",
    "structlog>=24.0.0",
    "httpx>=0.25.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends.legacy:build"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
